refactor(tests): split scheduler_test.go into focused modules
Remove **/*_test.go from the file_size.py blanket exclusion so Go test
files are now subject to the 500-line size policy. The six existing
oversized Go test files are grandfathered (tracked for follow-up splits).

scheduler_test.go (2 295 lines) is split into eight focused files, each
under 500 lines:

- scheduler_helpers_test.go   — shared mocks, YAML fixtures, helpers
- scheduler_lifecycle_test.go — core e2e and run-lifecycle tests
- scheduler_execution_test.go — workflow execution and template tests
- scheduler_state_transitions_test.go — step state transitions, error paths
- scheduler_step_limits_test.go — resolveStepLimits cascade tests
- scheduler_budget_test.go    — budget enforcement and token recording
- scheduler_cost_reporting_test.go — cost reporting and PR #86 tests
- scheduler_metadata_test.go  — step execution metadata tests
