+++ FILEMAP.md
@@ -1,13 +1,13 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-04-25  
-> **406** tracked files
+> **413** tracked files
 
 ## Summary
 
 | Top-level directory | Files |
 |---------------------|-------|
-| `internal/` | 93 |
+| `internal/` | 100 |
 | `docs/` | 81 |
 | `agents/` | 77 |
 | `tests/` | 65 |
@@ -28,7 +28,7 @@
 |-----------|-------|
 | `.py` | 153 |
 | `.md` | 97 |
-| `.go` | 90 |
+| `.go` | 97 |
 | `.yaml` | 25 |
 | `.rs` | 9 |
 | `.yml` | 6 |
@@ -396,7 +396,14 @@
 │   ├── scheduler/
 │   │   ├── budget.go
 │   │   ├── scheduler.go
-│   │   ├── scheduler_test.go
+│   │   ├── scheduler_budget_test.go
+│   │   ├── scheduler_cost_reporting_test.go
+│   │   ├── scheduler_execution_test.go
+│   │   ├── scheduler_helpers_test.go
+│   │   ├── scheduler_lifecycle_test.go
+│   │   ├── scheduler_metadata_test.go
+│   │   ├── scheduler_state_transitions_test.go
+│   │   ├── scheduler_step_limits_test.go
 │   │   └── stage_runner.go
 │   ├── security/
 │   │   └── security.go
